+++ CHANGELOG.md
@@ -5,6 +5,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- SqlServerDsc
+  - Added new public command:
+    - `Get-SqlDscConfigurationOption` - Returns the available configuration
+      options that can be used with the DSC resource _SqlConfiguration_.
+
 ## [16.3.0] - 2023-04-26
 
 ### Remove